+++ network/newstream/syncing_test.go
@@ -520,7 +520,7 @@ func getChunkCount(store chunk.Store) (c uint64, err error) {
 func getChunks(store chunk.Store) (chunks map[string]struct{}, err error) {
 	chunks = make(map[string]struct{})
 	for po := uint8(0); po <= chunk.MaxPO; po++ {
-		last, err := store.LastPullSubscriptionBinID(uint8(po))
+		last, err := store.LastPullSubscriptionBinID(po)
 		if err != nil {
 			return nil, err
 		}
